# lib/api-zod/src/generated/types/chatMessageInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WWW Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface ChatMessageInput {
  message: string;
  /** @nullable */
  projectId?: string | null;
  /** @nullable */
  context?: string | null;
}