# lib/api-zod/src/generated/types/chatResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WWW Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface ChatResponse {
  reply: string;
  messageId: string;
  /** @nullable */
  codeChanges?: string | null;
  suggestions?: string[];
}