# lib/api-zod/src/generated/types/cloneResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WWW Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface CloneResult {
  projectId: string;
  componentTree: string;
  thumbnailUrl: string;
  title: string;
  detectedStyle?: string;
  colors?: string[];
}