# lib/api-zod/src/generated/types/galleryStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WWW Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { GalleryStatsByStyle } from './galleryStatsByStyle';

export interface GalleryStats {
  total: number;
  byStyle: GalleryStatsByStyle;
  featured: number;
}