# lib/api-zod/src/generated/types/mobileTokenExchangeRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WWW Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface MobileTokenExchangeRequest {
  code: string;
  code_verifier: string;
  redirect_uri: string;
  state: string;
  nonce?: string;
}