# lib/api-zod/src/generated/types/theme.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WWW Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface Theme {
  id: string;
  name: string;
  description: string;
  preview: string;
}