# lib/api-zod/src/generated/types/themeInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WWW Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface ThemeInput {
  projectId: string;
  /** @nullable */
  themePreset?: string | null;
  /** @nullable */
  themePrompt?: string | null;
}