# lib/api-zod/src/generated/types/themeResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WWW Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { ThemeResultCssVariables } from './themeResultCssVariables';

export interface ThemeResult {
  themeTokens: string;
  presetName: string;
  cssVariables?: ThemeResultCssVariables;
}